pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "etool"
version = "1.3.2"
authors = [
  { name="Allen", email="jiangyangcreate@gmail.com" },
]
description = "A etool package for security monitoring"
readme = "README.md"
requires-python = ">=3.6"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "schedule>=1.1.0",
    "speedtest-cli>=2.1.3",
    "pynvml>=11.4.1",
    "flask",
    "pillow",
    "PyPDF2",
    "openpyxl",
    "numpy",
    "numba",
    "opencv-python-headless",
    "docx",
    "easyqr",
    "MyQR",
    "qrcode",
]

[project.urls]
Homepage = "https://github.com/jiangyangcreate/etool"
Issues = "https://github.com/jiangyangcreate/etool/issues"